backend/pyproject.toml
[project]
name = "fashion-ai-backend"
version = "0.1.0"
description = "FastAPI backend scaffold for fashion inspiration AI."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "openai>=1.0.0",
  "faiss-cpu>=1.8.0",
  "numpy>=1.26.0",
  "pillow>=10.0.0",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
